+++ CHANGELOG.md
@@ -10,6 +10,8 @@ these changes are prefixed with "**BREAKING**"
 ### Added
 
 * `requirements.txt` mainly to try and help GitHub detect dependencies
+* `delphin.derivation.UDFNode.parent` ([#245][])
+* `delphin.derivation.UDFTerminal.parent` ([#245][])
 
 ### Changed
 
@@ -1196,6 +1198,7 @@ information about changes, except for
 [#200]: https://github.com/delph-in/pydelphin/issues/200
 [#203]: https://github.com/delph-in/pydelphin/issues/203
 [#213]: https://github.com/delph-in/pydelphin/issues/213
+[#245]: https://github.com/delph-in/pydelphin/issues/245
 [#247]: https://github.com/delph-in/pydelphin/issues/247
 [#248]: https://github.com/delph-in/pydelphin/issues/248
 [#249]: https://github.com/delph-in/pydelphin/issues/249